toYocto added

+++ src/simple/assembly/model.ts
@@ -1,4 +1,4 @@
-import { PersistentUnorderedMap, PersistentVector, context } from "near-sdk-as";
+import { PersistentUnorderedMap, PersistentVector, context, u128 } from "near-sdk-as";
 export const payStatus = new PersistentUnorderedMap<string, Payment>("pay");
 export let studentInfo = new PersistentVector<Student>("info");
 export const day30= 2592000000000000
@@ -32,11 +32,11 @@ export class Payment {
  
   _status: bool;
   _payMount: u64;
-  _count: u64;
+  _count: u128;
   constructor(
     status: bool,
     payMount:u64,
-    count:u64,
+    count:u128,
 
   ) {
     this._status = status;
